+++ docs_src/how_to_guides/hashing_inputs_of_tasks_example_3_py310.py
@@ -9,7 +9,7 @@ from pytask import Product
 from pytask import PythonNode
 
 
-def calculate_hash(x: Any) -> str:
+def calculate_hash(x: Any) -> int | str:
     return DeepHash(x)[x]